REV6
The problem with this revision is that for some reason when the GUI is launched and there is less than 36 lines of data in the log file then the graph in the GUI will display N2O data only regardless if the graph is switched to a different sensor. The workaround for this is to comment out lines 162 and 163 in GUI_GRAPHING. The problem with this is that it doesn't update the N2O plot when new data arrives unless the N2O graph button is clicked again. However, the other two graphs (temp and humidity) update just fine.